[refactor] delete duplicated, change VideoPreviewResonse location
