+++ backend/src/main/java/com/woowacourse/edd/application/converter/VideoConverter.java
@@ -1,6 +1,6 @@
 package com.woowacourse.edd.application.converter;
 
-import com.woowacourse.edd.application.dto.VideoPreviewResponse;
+import com.woowacourse.edd.application.response.VideoPreviewResponse;
 import com.woowacourse.edd.application.dto.VideoSaveRequestDto;
 import com.woowacourse.edd.application.response.VideoResponse;
 import com.woowacourse.edd.domain.Video;
