+++ backend/src/main/java/com/woowacourse/edd/presentation/controller/VideoController.java
@@ -1,6 +1,6 @@
 package com.woowacourse.edd.presentation.controller;
 
-import com.woowacourse.edd.application.dto.VideoPreviewResponse;
+import com.woowacourse.edd.application.response.VideoPreviewResponse;
 import com.woowacourse.edd.application.dto.VideoSaveRequestDto;
 import com.woowacourse.edd.application.response.VideoResponse;
 import com.woowacourse.edd.application.service.VideoService;
